Respect outfile
I'm not sure what Myrddin's contributing guidelines are, so I'm doing
things the hg-git way of sending the patch to the mailing list.

This is really simple: it just actually uses outfile. If someone specifies
an output file, then it will be used as the output for the object file; the
output's basename will also be used for the use file and the assembler file
(if -S was passed).

Another patch is attached that fixes a slight logic error in the man page
that confused me very much. :)