Resolve trait decl types before fixing them
Otherwise, if the decl has a named type with parameters, it remains
unresolved until tyfix. tyfix looks up the tid in tytab, finding the
original definition of the named type, which may have different type
parameter names than than used in the trait decl.

Upon importing this trait from a use file and writing a new impl, we end
up trying to resolve type parameters of the *definition* of the named
type, rather than what is specified in the trait decl type. If both type
parameters are the same, we don't notice a problem, but if they differ,
we end up trying to serialize a Tyvar in genericname.

	tyvar has no idstr