Test exp.
We use Tang's implementation, which occasionally is off by 1 ulp.
In the double-precision case, this is about par with current libc
implementations. Glibc computes expf() in double-precision, so is
always (at least, as observed) last-place accurate. We appear to
be about on par with musl-libc in the single-precision case.

In order to increase accuracy, we could either decrease the interval
size (and increase table size), or switch to a different algorithm.
Decreasing the interval size is non-trivial, however, as it would
require recomputing the entire table, which is non-trivial.